backend/requirements.txt
# ============================================================
# Energy Project Intelligence Engine - Backend Dependencies
# ============================================================

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP & Scraping
httpx==0.27.0
requests==2.32.2
beautifulsoup4==4.12.3
lxml==5.2.1
selenium==4.21.0
webdriver-manager==4.0.1
playwright==1.44.0

# PDF Processing
PyMuPDF==1.24.3
pdfplumber==0.11.1
pypdf==4.2.0

# AI & LLM
openai==1.30.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
faiss-cpu==1.8.0
tiktoken==0.7.0

# ML
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2

# Geolocation
geopy==2.4.1

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
celery==5.4.0
redis==5.0.4
tenacity==8.3.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging
structlog==24.2.0
rich==13.7.1

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
